+++ influxdb_test.go
@@ -1,47 +1,75 @@
 package influxdb_test
 
 import (
-	"encoding/json"
+	"reflect"
 	"testing"
+	"time"
 
 	"github.com/influxdb/influxdb"
+	"github.com/influxdb/influxdb/client"
 )
 
-// Ensure that data with epoch timestamps can be decoded.
-func TestBatchPoints_Normal(t *testing.T) {
-	var p influxdb.BatchPoints
-	data := []byte(`
-{                                                                                                                                                       
-    "database": "foo",                                                                                                                                    
-    "retentionPolicy": "bar",                                                                                                                              
-    "points": [                                                                                                                                            
-        {                                                                                                                                                   
-            "name": "cpu",                                                                                                                                   
-            "tags": {                                                                                                                                         
-                "host": "server01"                                                                                                                            
-            },
-            "timestamp": 14244733039069373,
-            "precision": "n",
-            "values": {
-                    "value": 4541770385657154000
-            }
-        },
-        {
-            "name": "cpu",
-             "tags": {
-                "host": "server01"
-            },
-            "timestamp": 14244733039069380,
-            "precision": "n",
-            "values": {
-                    "value": 7199311900554737000
-            }
-        }
-    ]
-}
-`)
+func TestNormalizeBatchPoints(t *testing.T) {
+	now := time.Now()
+	tests := []struct {
+		name string
+		bp   client.BatchPoints
+		p    []influxdb.Point
+		err  string
+	}{
+		{
+			name: "default",
+			bp: client.BatchPoints{
+				Points: []client.Point{
+					{Name: "cpu", Tags: map[string]string{"region": "useast"}, Timestamp: now, Fields: map[string]interface{}{"value": 1.0}},
+				},
+			},
+			p: []influxdb.Point{
+				{Name: "cpu", Tags: map[string]string{"region": "useast"}, Timestamp: now, Fields: map[string]interface{}{"value": 1.0}},
+			},
+		},
+		{
+			name: "merge timestamp",
+			bp: client.BatchPoints{
+				Timestamp: now,
+				Points: []client.Point{
+					{Name: "cpu", Tags: map[string]string{"region": "useast"}, Fields: map[string]interface{}{"value": 1.0}},
+				},
+			},
+			p: []influxdb.Point{
+				{Name: "cpu", Tags: map[string]string{"region": "useast"}, Timestamp: now, Fields: map[string]interface{}{"value": 1.0}},
+			},
+		},
+		{
+			name: "merge tags",
+			bp: client.BatchPoints{
+				Tags: map[string]string{"day": "monday"},
+				Points: []client.Point{
+					{Name: "cpu", Tags: map[string]string{"region": "useast"}, Timestamp: now, Fields: map[string]interface{}{"value": 1.0}},
+					{Name: "memory", Timestamp: now, Fields: map[string]interface{}{"value": 2.0}},
+				},
+			},
+			p: []influxdb.Point{
+				{Name: "cpu", Tags: map[string]string{"day": "monday", "region": "useast"}, Timestamp: now, Fields: map[string]interface{}{"value": 1.0}},
+				{Name: "memory", Tags: map[string]string{"day": "monday"}, Timestamp: now, Fields: map[string]interface{}{"value": 2.0}},
+			},
+		},
+	}
 
-	if err := json.Unmarshal(data, &p); err != nil {
-		t.Errorf("failed to unmarshal nanosecond data: %s", err.Error())
+	for _, test := range tests {
+		t.Logf("running test %q", test.name)
+		p, e := influxdb.NormalizeBatchPoints(test.bp)
+		if test.err == "" && e != nil {
+			t.Errorf("unexpected error %v", e)
+		} else if test.err != "" && e == nil {
+			t.Errorf("expected error %s, got <nil>", test.err)
+		} else if e != nil && test.err != e.Error() {
+			t.Errorf("unexpected error. expected: %s, got %v", test.err, e)
+		}
+		if !reflect.DeepEqual(p, test.p) {
+			t.Logf("expected: %+v", test.p)
+			t.Logf("got:      %+v", p)
+			t.Error("failed to normalize.")
+		}
 	}
 }