+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - [#1864](https://github.com/influxdb/influxdb/pull/1864): fix race in startStateLoop
 - [#1753](https://github.com/influxdb/influxdb/pull/1874): Do Not Panic on Missing Dirs
 - [#1877](https://github.com/influxdb/influxdb/pull/1877): Broker clients track broker leader
-- [#1862](https://github.com/influxdb/influxdb/pull/1862): Fix memory leak in `httpd.serveWait`
+- [#1862](https://github.com/influxdb/influxdb/pull/1862): Fix memory leak in `httpd.serveWait`. Thanks @mountkin
 
 ## v0.9.0-rc9 [2015-03-06]